style: moving style related event to other util file
